requirements.txt
alembic==1.7.4
atomicwrites==1.4.0
attrs==21.2.0
click==8.0.3
colorama==0.4.4
concurrent-log-handler==0.9.19
dataclasses==0.8
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.8.1
importlib-resources==5.3.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
packaging==21.0
pip-autoremove==0.10.0
pluggy==1.0.0
portalocker==2.3.2
py==1.10.0
PyMySQL==1.0.2
pyparsing==2.4.7
pypiwin32==223
pytest==6.2.5
pytest-dotenv==0.5.2
pytest-mock==3.6.1
python-dotenv==0.19.1
pywin32==302
six==1.16.0
SQLAlchemy==1.4.25
toml==0.10.2
typing-extensions==3.10.0.2
Werkzeug==2.0.2
zipp==3.6.0
